UPI Fraud/requirements.txt
flask>=3.0.0
numpy
pandas
werkzeug
flask_mysqldb
sockets
mysql-connector-python
scikit-learn==1.3.2
matplotlib
seaborn